change to float(utils and bruteforce)

+++ bruteforce/BruteForceImplementation.cpp
@@ -1,9 +1,9 @@
 #include "BruteForceImplementation.h"
 #include "../utils_functions/Euclidean_dist.h"
 
-vector <pair<double,int>> brute_force_search(const vector<vector<double>> &pts, vector<double> q, int N){
+vector <pair<float,int>> brute_force_search(const vector<vector<float>> &pts, vector<float> q, int N){
     // collect the distances for all the queries and get the N top
-    vector<pair<double,int>> all;
+    vector<pair<float,int>> all;
     all.reserve(pts.size());
     double dist = 0.0;
     for (int i = 0 ; i < static_cast<int>(pts.size()) ; i++) {
